rename / make build work on linux
